Move message collecting topic consumer to separate class MODINVSTOR-1001
