+++ src/test/java/org/folio/rest/support/kafka/FakeKafkaConsumer.java
@@ -1,25 +1,14 @@
 package org.folio.rest.support.kafka;
 
-import static java.util.Collections.emptyList;
-
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.folio.kafka.services.KafkaEnvironmentProperties;
 import org.folio.rest.support.messages.EventMessage;
 
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
-import io.vertx.kafka.client.consumer.KafkaConsumer;
 import io.vertx.kafka.client.consumer.KafkaConsumerRecord;
-import io.vertx.kafka.client.serialization.JsonObjectDeserializer;
 
 public final class FakeKafkaConsumer {
   // These definitions are deliberately separate to the production definitions
@@ -73,7 +62,7 @@ public final class FakeKafkaConsumer {
   }
 
   public static Collection<EventMessage> getMessagesForAuthority(String authorityId) {
-    return authorityTopicConsumer.receivedMessages(authorityId);
+    return authorityTopicConsumer.receivedMessagesByKey(authorityId);
   }
 
   public static int getAllPublishedInstanceIdsCount() {
@@ -81,7 +70,7 @@ public final class FakeKafkaConsumer {
   }
 
   public static Collection<EventMessage> getMessagesForInstance(String instanceId) {
-    return instanceTopicConsumer.receivedMessages(instanceId);
+    return instanceTopicConsumer.receivedMessagesByKey(instanceId);
   }
 
   public static Collection<EventMessage> getMessagesForInstances(List<String> instanceIds) {
@@ -94,17 +83,17 @@ public final class FakeKafkaConsumer {
   public static Collection<EventMessage> getMessagesForHoldings(
     String instanceId, String holdingsId) {
 
-    return holdingsTopicConsumer.receivedMessages(instanceAndIdKey(instanceId, holdingsId));
+    return holdingsTopicConsumer.receivedMessagesByKey(instanceAndIdKey(instanceId, holdingsId));
   }
 
   public static Collection<EventMessage> getMessagesForItem(
     String instanceId, String itemId) {
 
-    return itemTopicConsumer.receivedMessages(instanceAndIdKey(instanceId, itemId));
+    return itemTopicConsumer.receivedMessagesByKey(instanceAndIdKey(instanceId, itemId));
   }
 
   public static Collection<EventMessage> getMessagesForBoundWith(String instanceId) {
-    return boundWithTopicConsumer.receivedMessages(instanceId);
+    return boundWithTopicConsumer.receivedMessagesByKey(instanceId);
   }
 
   private static String instanceAndIdKey(String instanceId, String itemId) {
@@ -119,64 +108,5 @@ public final class FakeKafkaConsumer {
     final var id = oldOrNew != null ? oldOrNew.getString("id") : null;
 
     return instanceAndIdKey(message.key(), id);
-  }
-
-  public static class MessageCollectingTopicConsumer {
-    private final String topicName;
-    private final Map<String, List<EventMessage>> collectedMessages;
-    private final Function<KafkaConsumerRecord<String, JsonObject>, String> keyMap;
-    private KafkaConsumer<String, JsonObject> consumer;
-
-    public MessageCollectingTopicConsumer(String topicName,
-      Function<KafkaConsumerRecord<String, JsonObject>, String> keyMap) {
-
-      this.topicName = topicName;
-      this.collectedMessages = new ConcurrentHashMap<>();
-      this.keyMap = keyMap;
-    }
-
-    private void subscribe(Vertx vertx) {
-      consumer = KafkaConsumer.create(vertx, consumerProperties());
-
-      consumer.handler(this::acceptMessage);
-      consumer.subscribe(topicName);
-    }
-
-    public void unsubscribe() {
-      if (consumer != null) {
-        consumer.unsubscribe();
-      }
-    }
-
-    private void acceptMessage(KafkaConsumerRecord<String, JsonObject> message) {
-      final var collectedMessages = this.collectedMessages.computeIfAbsent(
-        this.keyMap.apply(message), v -> new ArrayList<>());
-
-      collectedMessages.add(EventMessage.fromConsumerRecord(message));
-    }
-
-    private Collection<EventMessage> receivedMessages(String key) {
-      return collectedMessages.getOrDefault(key, emptyList());
-    }
-
-    private int countOfReceivedKeys() {
-      return collectedMessages.size();
-    }
-
-    private void discardCollectedMessages() {
-      collectedMessages.clear();
-    }
-
-    private static Map<String, String> consumerProperties() {
-      Map<String, String> config = new HashMap<>();
-      config.put("bootstrap.servers", KafkaEnvironmentProperties.host() + ":" + KafkaEnvironmentProperties.port());
-      config.put("key.deserializer", StringDeserializer.class.getName());
-      config.put("value.deserializer", JsonObjectDeserializer.class.getName());
-      config.put("group.id", "folio_test");
-      config.put("auto.offset.reset", "earliest");
-      config.put("enable.auto.commit", "false");
-
-      return config;
-    }
   }
 }
